"Chase camera" image cropper
Process only the area nearby previously detected objects, since we know that
the user can only move his head so fast, and we want to reduce the number of
pixels we have to process each frame to keep CPU usage down and frame rate
up.

We pass coordinates of cropped frame back to the tracker so that it
can (a) normalize its detected object positions and (b) draw the cropped
frame on the visualizer.

Add logging config to headMouse.py.