Visualize detected objects as squares, not circles
Draw rectangles around detected objects per the bounding boxes we get
from the cascade trackers. This corresponds better to the actual boundaries
we have.

This also fixes a bug in computing the radius of the circle we were drawing
around the detected objects, which was previously twice as large as it
should have been.

Change the cursor output visualization to be a smaller, white circle with
a thicker trace. This is easier to see than the bigger, thinner, blue circle.