Add search area scaling to eye tracker
If our search area window crop returns a large area, scale to get closer to about
10–15 kilopixels per eye.

We achieve this by scaling and search area on one or two objects to 100 px high.
If the scale factor is >= 0.75, i.e. if it would only moderately reduce the size
of the search area, we skip the scaling entirely. Similarly, if the scale factor
would increase the size of the search area, adding pixels without adding actual
input data, we skip it.

Otherwise, we scale the search area to 100 px high and rescale the found object
output coordinates after the cascade tracker runs.

Additionally, we're now (temporarily) cropping the object list to one item. Ideally,
we'd pick the best object, but cropping to one seems to work well enough for now.

By cropping to one, we ensure we use scaling whenever it's appropriate, keeping us
around 10 kilopixels per search regardless of head position.

The biggest win is when the face is close to the camera, making the eye take up
many pixels with little benefit.

Finally, we're logging scale factor stats alongside the search area kpx stats.