Load eye cascade XML file from package_data
Include the default eye cascade detector XML as package_data
so we're not dependent on the CWD to load the file. This
should in future be overridable with a configuration.

Remove the face cascade input driver since it's not implemented
and really should be just an option to change the cascade XML
for the "eye" tracker driver.