Add configuration for distance_scaling
Allow turning distance scaling on and off (on by default) from the
config system.

Note that since the distance scaling radically alters the reported
input velocities, the acceleration and sensitivity params will
need to be adjusted as scaling is turned on and off.

Fixing this should be part of a unified patch to (a) scale distances
to some standardized value that's compatible with having scaling
turned off and (b) scale input driver (x, y) coordinates to be camera
resolution independent.